refactor: 클래스 파일 패키지 이동 & 불필요한 변수 삭제
