+++ src/main/java/util/HttpRequestUtil.java
@@ -1,6 +1,6 @@
 package util;
 
-import http.response.ContentType;
+import http.enums.ContentType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
